# Conflicts:
#	package.json
#	projects/angular-highcharts-declarative/package.json
#	projects/angular-highcharts-declarative/src/lib/children/series/hc-polygon.component.ts

+++ projects/angular-highcharts-declarative/src/lib/children/hc-series.component.ts
@@ -23,19 +23,19 @@ import {
   SeriesMapDataOptions,
   SeriesOptions
 } from 'highcharts';
-import {HcChartService} from '../hc-chart.service';
-import {BehaviorSubject, Observable, Subscription} from 'rxjs';
-import {first} from 'rxjs/operators';
-import {changesToFlat, registerEvents} from '../helpers';
-import {HcTooltipComponent} from './hc-tooltip.component';
-import {SeriesLegendItemClickEventObject} from 'highcharts/highcharts.src';
-import {HcPointComponent} from './hc-point.component';
+import { HcChartService } from '../hc-chart.service';
+import { BehaviorSubject, Observable, Subscription } from 'rxjs';
+import { first } from 'rxjs/operators';
+import { changesToFlat, registerEvents } from '../helpers';
+import { HcTooltipComponent } from './hc-tooltip.component';
+import { SeriesLegendItemClickEventObject } from 'highcharts/highcharts.src';
+import { HcPointComponent } from './hc-point.component';
 
 @Component({
   selector: 'hc-series',
   template: `
-    <ng-content select="hc-tooltip"></ng-content>
-    <ng-content select="hc-point"></ng-content>
+      <ng-content select="hc-tooltip"></ng-content>
+      <ng-content select="hc-point"></ng-content>
   `,
   changeDetection: ChangeDetectionStrategy.OnPush
 })
@@ -92,8 +92,10 @@ export class HcSeriesComponent implements OnInit, OnDestroy, OnChanges, SeriesOp
   initialized$ = this.initializedSub.pipe(first(v => v));
   private dataSub: Subscription;
   private changesSub: Subscription;
+  private _rawSeries: Series;
 
-  constructor(private chartService: HcChartService, private zone: NgZone) {}
+  constructor(private chartService: HcChartService, private zone: NgZone) {
+  }
 
   ngOnInit() {
     this.initialized$.subscribe(() => {
@@ -124,7 +126,8 @@ export class HcSeriesComponent implements OnInit, OnDestroy, OnChanges, SeriesOp
     this.index = this.index || index;
     this.chartService.addSeries(
       { ...this.getState(), ...{ events: registerEvents(this, this.zone, 'series') } },
-      () => {
+      (series: Series) => {
+        this._rawSeries = series;
         this.changesSub = this.tooltips.changes.subscribe(v => this.tooltips.forEach(t => t.setSeries(this.index)));
         this.tooltips.forEach(t => t.setSeries(this.index));
         this.initializedSub.next(true);
@@ -143,6 +146,7 @@ export class HcSeriesComponent implements OnInit, OnDestroy, OnChanges, SeriesOp
     delete state.dataSub;
     delete state.zone;
     delete state.hcPoint;
+    delete state._rawSeries;
     state = { ...state, ...state.extra };
     for (const [key, value] of Object.entries(state)) {
       if (value === null) {
@@ -153,11 +157,18 @@ export class HcSeriesComponent implements OnInit, OnDestroy, OnChanges, SeriesOp
   }
 
   addPoint(value: any, shift = this.dataStreamShift) {
-    this.chartService.addPoint(this.index, value, shift);
+    this._rawSeries.addPoint(value, true, shift);
+    // this.chartService.addPoint(this.index, value, shift);
   }
 
   update(props: Partial<SeriesOptions>) {
-    this.chartService.updateSeries(this.index, props);
+    this._rawSeries.update(props as any);
+    // this.chartService.updateSeries(this.index, props);
+  }
+
+  updateSeriesData(data: any) {
+    this._rawSeries.setData(data);
+    // this.chartService.updateSeriesData(this.index, changes.data.currentValue);
   }
 
   ngOnChanges(changes: SimpleChanges) {
@@ -165,9 +176,8 @@ export class HcSeriesComponent implements OnInit, OnDestroy, OnChanges, SeriesOp
       return;
     }
     if (changes.data && !changes.data.isFirstChange()) {
-      this.chartService.updateSeriesData(this.index, changes.data.currentValue);
+      this.updateSeriesData(changes.data.currentValue);
     }
-
     if (changes.dataStream && !changes.dataStream.isFirstChange()) {
       if (this.dataSub) {
         this.dataSub.unsubscribe();
@@ -187,6 +197,15 @@ export class HcSeriesComponent implements OnInit, OnDestroy, OnChanges, SeriesOp
     if (this.changesSub) {
       this.changesSub.unsubscribe();
     }
-    this.chartService.removeSeries(this.index);
+    if (this._rawSeries) {
+      if (typeof (this._rawSeries.remove) === 'function') {
+        this._rawSeries.remove();
+      } else {
+        this.chartService.removeSeries(this.index);
+      }
+      this._rawSeries = null;
+    }
+
+    // this.chartService.removeSeries(this.index);
   }
 }
